+++ source/control.lua
@@ -1,15 +1,12 @@
 
 modVersion = "0.3.11"
 
--- Tables used for registering common events of entities
--- [$entityName] = { build = $function(entity), tick = $function(entity,data), remove = $function(data) }
-entities = {}
-
-
 require "defines"
 require "libs.functions"
 require "libs.controlFunctions"
-require "libs.gui"
+
+require "libs.entities" --lets your classes register event functions in general
+require "libs.gui" --lets you register functions when a entity gui is opened/closed
 
 require "control.belt-sorter"
 require "control.incinerators"
@@ -20,13 +17,7 @@ require "control.migration_0_3_11"
 -- hardCrafting.version = $version
 -- hardCrafting.incinerators = { $incinerator:LuaEntity, ... }
 -- hardCrafting.eincinerators = { $incinerator:LuaEntity, ... }
-
--- Currently only belt-sorters use generic setup:
--- hardCrafting.schedule[tick][idEntity] = $entity
--- hardCrafting.entityData[idEntity] = { name=$name, ... }
-
--- Constants:
-TICK_ASAP = 0 --game.tick used in migration when game variable is not available yet
+-- Currently only belt-sorters use generic entity system
 
 -- Init --
 script.on_init(function()
@@ -47,100 +38,23 @@ function init()
 	if hc.schedule == nil then hc.schedule = {} end
 	if hc.entityData == nil then hc.entityData = {} end
 	if hc.playerData == nil then hc.playerData = {} end
-
+	info(hc.version)
 	if hc.version < "0.3.11" then migration_0_3_11() end
-	initGui()
+	
+	entities_init()
 end
+
+script.on_event(defines.events.on_gui_click, function(event)
+	local element = event.element
+	warn(element.name)
+end)
 
 script.on_event(defines.events.on_tick, function(event)
 	tickGui()
-	local hc = global.hardCrafting
 	updateIncinerators()
-
-	-- schedule events from migration
-	if hc.schedule[TICK_ASAP] ~= nil then
-		for id,entity in pairs(hc.schedule[TICK_ASAP]) do
-			scheduleAdd(entity, game.tick )-- TODO: use randomized distribution math.random(8))
-		end
-		hc.schedule[TICK_ASAP] = nil
-	end
-
-	-- if no updates are scheduled return
-	if hc.schedule[game.tick] == nil then
-		return
-	end
-	--dummyUpdate()
-	-- --[[
-	-- Execute all scheduled events
-	for entityId,entity in pairs(hc.schedule[game.tick]) do
-		if entity and entity.valid then
-			local data = hc.entityData[idOfEntity(entity)]
-			local name = entity.name
-			local nextUpdateInXTicks, reasonMessage
-			if entities[name] ~= nil then
-				if entities[name].tick ~= nil then
-					nextUpdateInXTicks, reasonMessage = entities[name].tick(entity,data)
-				end
-			else
-				warn("updating entity with unknown name: "..name)
-			end
-			if reasonMessage then
-				info(name.." at " .. entity.position.x .. ", " ..entity.position.y .. ": "..reasonMessage)
-			end
-			if nextUpdateInXTicks then
-				scheduleAdd(entity, game.tick + nextUpdateInXTicks)
-				-- if no more update is scheduled, remove it from memory
-				-- nothing to be done here, the entity will just not be scheduled anymore
-			end
-		elseif entityId == "text" then
-			PlayerPrint(entity)
-		else
-			-- if entity was removed, remove it from memory
-			local data = hc.entityData[entityId]
-			local name = data.name
-			if entities[name] ~= nil then
-				if entities[name].remove ~= nil then
-					entities[name].remove(data)
-				end
-			else
-				info("removing "..name.." at: "..entityId)
-			end
-			hc.entityData[entityId] = nil
-		end
-	end
-	-- ]]--
-	global.hardCrafting.schedule[game.tick] = nil
+	entities_tick()
 end)
 
-
-function dummyUpdate()
-	log("starting update")
-	local hc = global.hardCrafting
-	for entityId,entity in pairs(hc.schedule[game.tick]) do
-		local data = hc.entityData[entityId]
-		beltSorterSearchInputOutput(entity,data)
-	end
-	log("input/output search done")
-	for entityId,entity in pairs(hc.schedule[game.tick]) do
-		local data = hc.entityData[entityId]
-		beltSorterBuiltFilter(entity,data)
-	end
-	log("filter build done")
-	for entityId,entity in pairs(hc.schedule[game.tick]) do
-		local data = hc.entityData[entityId]
-		beltSorterDistributeItems(entity,data)
-	end
-	log("item distribute done")
-	for _,entity in pairs(hc.schedule[game.tick]) do
-		scheduleAdd(entity, game.tick + 8)
-	end
-	log("scheduling done")
-	local count = 0
-	for _,_ in pairs(hc.schedule[game.tick]) do
-		count = count +1
-	end
-	log("done with "..tostring(count).." belt-sorter")
-end
 
 ---------------------------------------------------
 -- Building Entities
@@ -153,23 +67,7 @@ script.on_event(defines.events.on_robot_built_entity, function(event)
 end)
 
 function entityBuilt(event)
-	local entity = event.created_entity
-	local name = entity.name
-	if entities[name] == nil then return end
-	global.hardCrafting.entityData[idOfEntity(entity)] = { ["name"] = name }
-	if entities[name].build then
-		local data = entities[name].build(entity)
-		table.addTable(global.hardCrafting.entityData[idOfEntity(entity)],data)
-	end
+	entities_build(event)
 end
 
----------------------------------------------------
--- Utility methods
----------------------------------------------------
--- Adds new entry to the scheduling table
-function scheduleAdd(entity, nextTick)
-	if global.hardCrafting.schedule[nextTick] == nil then
-		global.hardCrafting.schedule[nextTick] = {}
-	end
-	global.hardCrafting.schedule[nextTick][idOfEntity(entity)]=entity
-end
+
